+++ druid/src/lib.rs
@@ -120,9 +120,13 @@
 //! [`usvg` crate]: https://crates.io/crates/usvg
 //! [`image` crate]: https://crates.io/crates/image
 
-#![deny(intra_doc_link_resolution_failure, unsafe_code)]
+#![deny(
+    intra_doc_link_resolution_failure,
+    unsafe_code,
+    clippy::trivially_copy_pass_by_ref
+)]
+#![warn(missing_docs)]
 #![allow(clippy::new_ret_no_self, clippy::needless_doctest_main)]
-#![deny(clippy::trivially_copy_pass_by_ref)]
 #![cfg_attr(docsrs, feature(doc_cfg))]
 
 // Allows to use macros from druid_derive in this crate
@@ -182,7 +186,7 @@ pub use app_delegate::{AppDelegate, DelegateCtx};
 pub use box_constraints::BoxConstraints;
 pub use command::{sys as commands, Command, Selector, SingleUse, Target};
 pub use contexts::{EventCtx, LayoutCtx, LifeCycleCtx, PaintCtx, UpdateCtx};
-pub use data::Data;
+pub use data::{ArcStr, Data};
 pub use env::{Env, Key, KeyOrValue, Value, ValueType};
 pub use event::{Event, InternalEvent, InternalLifeCycle, LifeCycle};
 pub use ext_event::{ExtEventError, ExtEventSink};
@@ -210,4 +214,5 @@ pub(crate) use event::{StateCell, StateCheckFn};
 pub type KeyCode = KbKey;
 
 #[deprecated(since = "0.7.0", note = "Use Modifiers instead")]
+/// See [`Modifiers`](struct.Modifiers.html).
 pub type KeyModifiers = Modifiers;